Basic working open ondemand role
This is the initial commit for the open ondemand ansible role. At this
point it can build and install open ondemand, but is not ready for
production installs. Along with it's instability, it also relies on the
OS versions of nodejs and ruby which must be fairly high, so won't work
on say, CentOS 7 just yet.