Remove tag always from include_vars
If OOD role is defined along with other playbooks and a tag is used to
invoke non OOD related tasks, the OOD playbook will throw an error about
ansible_distribution not being defined.

This seems to be due the `tag: always` on include_vars tasks for ansible
specific distributions. This commits removes it.